requirements.txt
pytesseract
Pillow
pandas
python-dateutil
python-dotenv
openai
streamlit
pyarrow
fastparquet
requests
gspread
google-auth
pytest
